requirements.txt
boto3==1.24.12
pandas==1.3.4
psycopg2-binary==2.9.3
python-binance==1.0.15
requests==2.27.1
streamlit==1.10.0



